Link to Manage Applications from user edit page
We've updated the various granting_permissions_tests to reflect the UI
change.

We've had to update the setup of the following tests so that they
continue to pass.

    should "support granting app-specific permissions"

We've had to grant the user access to the app as it's no longer possible
to give a user any permissions to an app unless they first have the
signin permission for that app.

    should "not be able to grant permissions that are not grantable_from_ui"

We have to add signin permission, and another permission that is
grantable from the UI, so that we can get to the permissions/edit page:
we don't show the "Update permissions for <app>" link if the signin
permission is the only grantable permission from the UI.

We've removed the following tests that no longer make sense given the
current implementation.

    should "not remove permissions the user has that the super organisation admin does not have"
    should "not remove permissions the user has that the super organisation admin cannot delegate"
    should "not remove permissions the user has that the organisation admin does not have"
    should "not remove permissions the user has that the organisation admin cannot delegate"

These tests were necessary because in the previous implementation
another user's applications permissions were updated in bulk through the
edit user page. User's permissions are now updated for a single
application at a time. In order for a Publishing Manager to be able to
edit another user's permissions for an application they need to have the
signin permission for that application, and the signin permission needs
to be delegatable. This logic is defined in the
`UserApplicationPermissionPolicy`.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>