Extract string attribute names in ApiUsersController
I'm planning to change the "edit user" page to remove the form field for
the user name and instead provide a link to the change user name page
provided by `Users::NamesController#edit`. This will mean we no longer
want to permit the `api_user[:name]` param for the
`ApiUsersController#update` action, but we will want to preserve it for
the `#create` action. Extracting this local variable will make that
simpler.

Note that I've also renamed the `params` argument to
`ApiUsersController#permitted_user_params` to avoid it shadowing the raw
params, because I'm going to need the latter to detect which is the
current action.