Simplify UsersController#update
This action is now only used as the target of the form on the
`/users/<id>/require_2sv` page and that can't (currently at least) fail
validation. I considered displaying errors on the require_2sv page and
rendering that if `UserUpdate#call` fails but it doesn't seem worth the
effort.