Introduce Current.user & Current.user_ip
The `Current` class inherits from `ActiveSupport::CurrentAttributes` [1]
which means that its singleton attribute accessors are
"thread-isolated". This means that we can set them in a new
`before_action` in `ApplicationController` and then make use of them
in model code without needing access to a controller. I'm hoping to use
this to simplify calls to `EventLog` methods and ultimately move a bunch
of functionality in `UserUpdate#call` into model callbacks.

[1]: https://api.rubyonrails.org/v7.0.8/classes/ActiveSupport/CurrentAttributes.html