Use Current.user_ip in EventLog.record_event
The `ip_address` option for this method is optional and not all
callsites were setting it. This means we only record the current user's
IP address in some cases. To this end I've repurposed the `ip_address`
option to be a boolean value indicating whether or not we want to record
the user's IP address.