Remove skip_update_user_permissions param from UsersController
This was here to handle the use of `UsersController#update` when it was
invoked from the `/users/<id>/require_2sv` page. The `require_2sv` page
contains a form with a single checkbox that allows the
`User#require_2sv` boolean to be toggled. If the checkbox is left
unchecked then the controller doesn't receive a `user` param and the
call to `params.require(:user)` fails. We're avoiding this problem by
adding a hidden field to the form that defaults `require_2sv` to `0`
which is the same way that Rails' `check_box` helper handles it[1].

[1]: https://api.rubyonrails.org/classes/ActionView/Helpers/FormBuilder.html#method-i-check_box