Use UserApplicationPermission.for
We found the use of `User#signin_permission_for` confusing because it
seems to imply that there's something special about this instance of
`UserApplicationPermission`. In fact, any instance of
`UserApplicationPermission` that links to both a `User` and
`Application` contains enough information for us to determine whether
the current user is authorised to edit a user's permissions for an
application. The use of `UserApplicationPermission.for` in this commit
attempts to make that clearer.

Ideally we'd retrofit this change into the earlier commits that use
`User#signin_permission_for` but we've decided that it's not worth the
effort.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>