Use Current.user in EventLog.record_event
The `initiator` option for this method is optional and not all callsites
were setting it. This means we only record the initiating user in some
cases. To this end I've repurposed the `initiator` option to be a
boolean value indicating whether or not we want to record the initiating
user.