Move UserUpdate#record_update -> User callback
This is now called *automatically* via an `after_update` callback which
means it's called more reliably and more frequently than
`UserUpdate#record_update` was called.

The `ActiveModel::Dirty#previous_changes` guard condition avoids
recording events when none of the user's attributes have changed which
would be pretty pointless.

The `Current.user` guard condition is necessary, because
`LogEvent::ACCOUNT_UPDATED` has `require_initiator` set to true and so
would trigger a validation error if `Current.user` was not set.

I've had to add a few stubs for extra `LogEvent::ACCOUNT_UPDATED` event
log records which reflect a slight change in behaviour. However, I don't
think these extra even log records are necessarily a bad thing!

The advantage of recording the event in a callback like this is that we
don't need to remember to use `UserUpdate#call` to make any changes to a
user. This in turn means that we will now be recording events when an
API user's attributes are updated. which we weren't previously because
the `ApiUsersController` wasn't calling `UserUpdate#call`.