WIP: Move UserUpdate#perform_permissions_update -> User callback
TODO: Previously this might have been called when only a user's
permissions had changed, i.e. none of its actual attributes, which would
mean that the `after_update` callback on `User` would not be called.

This is now called *automatically* via an `after_update` callback which
means it's called more reliably than
`UserUpdate#perform_permissions_update` was called.

I've had to add a call `#reload` on the `User#authorisations`
association in the test for `User#suspend`, because somehow the call to
`User#perform_permissions_update` was leaving the in-memory
`User#authorisations` out-of-date.

It's safe to remove a bunch of calls to `PermissionUpdater.perform_on`,
because in all cases the relevant user is being updated and so the new
callback will be fired.

The advantage of queueing the `PermissionUpdater` in a callback like
this is that we don't need to remember to use `UserUpdate#call` to make
any changes to a user and we may now catch some places where a user was
being updated, but the job was not being queued.

TODO: I've removed a test in `ApiUsersControllerTest`, because you can
no longer update name or email. However, it might be possible to get
it to work when updating permissions.