Move UserUpdate#record_role_change -> User callback
This is now called *automatically* via an `after_update` callback which
means it's called more reliably than `UserUpdate#record_role_change` was
called.

I've changed the implementation very slightly to make use of other magic
methods provided by `ActiveModel::Dirty#previous_changes`. I don't
believe any of these changes will have changed the behaviour.

The `Current.user` guard condition is necessary, because
`LogEvent::ROLE_CHANGED` has `require_initiator` set to true and so
would trigger a validation error if `Current.user` was not set.

The advantage of recording the event in a callback like this is that we
don't need to remember to use `UserUpdate#call` to make any changes to a
user and we may now catch some places where a user's role was being
updated but no event was being recorded.