Add ApiUsers::ApplicationsController#index
This page will allow a superadmin to see which applications an API
user has access to and eventually edit their permissions.

We've based the view on `/users/applications/index.html.erb` but as
managing access to applications is handled on the "manage tokens" page
we do not need to show applications the user does not have access to
here.

Note it is possible for api users to have access tokens for
applications that are not "api only" (e.g in production the "Content
Publisher" api user has a token for "Whitehall") so we do not filter
out applications based on their `api_only` field.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>