Prevent Admins from suspending/unsuspending API users
Previously, `SuspensionsController#load_and_authorize_user` was only
ever checking `UserPolicy#suspension?` to determine whether the current
user was authorized to suspend/unsuspend an API user. This meant that
although it hasn't been possible for Admins to *navigate* to the
suspend/unsuspend user page for an API user, it would have been easy
enough to hack the URL to reach the page and then there was nothing
preventing them from suspending/unsuspending the API user.

I think this was an oversight when the `UserPolicy#suspension?` was
introduced almost 9 years ago in #298. Since it was calling `User.find`
it would never have returned an instance of an `ApiUser` and thus it
would never have tried to use `ApiUserPolicy#suspension?` instead of
`UserPolicy#suspension?`. I've implemented the former by aliasing it to
`ApiUserPolicy#new?` as per all the other `ApiUser`-related predicates
which means that only Superadmins have permission.