Adapt Users::EmailsController authorization to work with API users
This ensures that we use predicate methods on an instance of the correct
policy class, i.e. `UserPolicy` or `ApiUserPolicy`. Previously it was
always using `UserPolicy` whether the user was an API user or not.

Also although it shouldn't be possible to trigger requests to
`Users::EmailsController#resend_email_change` or `#cancel_email_change`
through the UI, this adds predicate methods to `ApiUserPolicy` for these
actions to make them forbidden.

I do slightly wonder whether the latter behaviour is truly an
"authorization" concern, because really it's just that these operations
should not exist for an `ApiUser`. However, I think that's a bit
philosophical at this point.