Move status to details section on edit API user page
This is a small step towards moving this page to the latest design which
will use the GOV.UK Design System. I'm trying to make small changes to
the page which will make the transition easier while retaining all the
functionality and keeping all the tests passing.

I've used an HTML description list element, because this most closely
resembles the Summary List component [1] that we're planning to use.

Note that I'm using `UsersHelper#status` to format the user's status to
match the design.

c.f. this commit [2].

[1]: https://components.publishing.service.gov.uk/component-guide/summary_list
[2]: https://github.com/alphagov/signon/pull/2567/commits/109473965f5aabbfa786a3ac0efd58e71b14127b