Move edit API user page to use GOV.UK Design System 🎉
Trello: https://trello.com/c/bhPUU9qo

This introduces the Summary List component [1] that we've been working
towards for displaying the attributes of the user along with links to
change them. I've reused some of the `UsersHelper` methods that I
extracted in the equivalent commit [2] for the edit API user page to
generate the appropriate options for each of the items.

The links in the "Actions" section are displayed using a List
component [2] and its items are obtained from a bunch of
`UsersHelper#link_to_*` methods. As I mentioned when I added the latter,
these return `nil` if the link is not to be displayed and we handle
that here by compacting the array of items before passing it to the
component.

I've had to tweak a few of the assertions in the `ManageApiUsersTest`
integration test:
* The link text for the access log page has changed
* Flash alert messages are rendered differently in the new layout

[1]: https://components.publishing.service.gov.uk/component-guide/summary_list
[2]: https://github.com/alphagov/signon/pull/2567/commits/8d16e6fdcc05ad7a5e91ef977ff847a2e9aaed38
[3]: https://components.publishing.service.gov.uk/component-guide/list