Add ApiUsers::PermissionsController#edit/update
We've copied the controller, template and test files from their user
counterparts and modified as necessary. We've decided to add this
duplication for now but will see whether we can reduce/remove it later.

Note that we're using `UserUpdate` in this new controller, where we were
using `User#update` in `ApiUsersController#update`[1]. Using
`UserUpdate` has the following advantages:

- consistency with the other controllers that modify a user's
  permissions (account/permissions_controller and
  users/permissions_controller).
- record an `EventLog` about the update
- record an `EventLog` about the permissions that have been
  added/removed

[1]: https://github.com/alphagov/signon/blob/089516afdc0311525769d89c81bc837db4b16a2d/app/controllers/api_users_controller.rb#L42C25-L42C51

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>