Fix assertions for change links in edit API user page
The `href` option on these `assert_select` calls wasn't having any
effect, so they were only checking that a link with the given text was
present not where it was linking to!

Also there was a copy/paste error in the assertion for the "Change
Email" link - it was looking for the wrong URL path.