Refactor User#authorised_applications to use has_many
Previously this method was returning an array of unique applications
using Ruby's `group_by`. It's more idiomatic, and will allow us to
chain on additional ActiveRecord query methods if we convert it to a
has_many instead.

This method was only being used via its alias `:applications_used` so
we've changed the call site to use `:authorised_applications`
instead (which we think is clearer).

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>