Add routes for Users::NamesController under api_users
This means we can make `Users::NamesController#load_user` a bit more
idiomatic and it means the URL paths are a little more logical/hackable.

I haven't changed the form field name, i.e. it's always "user[name]"
rather than having it be "user[name]" for non-API users and
"api_user[name]" for API users, because I thought that was an
unnecessary complication, although obviously it's not so idiomatic.

Having the controller support two different sets of routes does make the
controller a bit more complicated, but I think it's probably worth it.
If it gets any more complicated, it might be worth splitting it up into
two controllers and removing the duplication some other way.