Avoid User#permitted_params calls from user-related controllers
These calls were a simpler way to retain the functionality provided by
`UserParameterSanitiser`. However, I've since realised that that
functionality was only really relevant in controller actions where we
were modifying more than one attribute of the user. In each of these
newer controllers we're only ever modifying a single attribute of a user
and so the authorization checks that we're doing in each controller are
enough to ensure the user is allowed to change the relevant attribute.

Unfortunately I've had to add a Brakeman warning to the ignore list. The
warning is about mass assignment of `User#role`. If the controller
wasn't using `UserUpdate` it would've been simple to solve this by not
doing a mass assigment, but since it is, that's quite awkward.