Fix issues in Users::PermissionsController#update
In addition to losing the signin permission (which we fixed in the
previous commit) we also noticed 3 other issues with the current
implementation. We attempted to add each of these tests in their own
commit but the fix we've come up with resolves all 3 so we've decided to
combine them.

1. should "not remove permissions the user already has that are not
   grantable from ui"

If a user had been granted a permission that isn't grantable from the UI
then submitting the form to update their permissions would result in
that permission being lost. In production the only permission not
grantable from the UI is the `user_update_permission` and this is only
given to the "Signon API Client" (in `SSOPushCredential.credentials`[1])
so we don't think this has been causing any real problems but we should
fix it nonetheless.

Note that we're sorting the `SupportedPermission#id`s that end up in the
arguments to `UserUpdate.new` to make it easier to test.

2. should "prevent permissions being added for other apps"

Although not possible through the user interface, it was possible to
craft a request to give permissions to applications other than the
application identified by `params[:application_id]` in the URL. There
was no real danger here because the `UserUpdate` service ensures that
the current user can only modify permissions for users and applications
they're allowed to manage.

3. should "prevent permissions being added that are not grantable from
   the ui"

Although not possible through the user interface, it was possible to
craft a request to add a permission that's not grantable through the UI.
As explained in (1) above, the only permission not grantable from the UI
is the `user_update_permission` and this doesn't have any effect for
users other than the "Signon API Client".

These same problems exists in `Account::PermissionsController#update` and
we intend to fix that in a separate commit.

[1]: https://github.com/alphagov/signon/blob/f6690d8d13890da3279313e47cedfcefaeeeb7e3/lib/sso_push_credential.rb#L11

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>