Retain signin permission for app when updating user permissions
We noticed that it was possible to construct a request to the
`Users::PermissionsController#update` action that resulted in the user's
signin permission being removed from the application. This won't happen
through normal use of the user interface but we should guard against it
nonetheless.

The same problem exists in `Account::PermissionsController#update` and
we intend to fix that in a separate commit.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>