Implement separate api_user_params for create & update actions
As suggested by @chrisroos. I agree that this makes it clearer what's
going on.

Also it has highlighted a potential security issue which is that the
form for the `create` action doesn't include permissions, but the
`create` action itself allows the `supported_permission_ids` param to be
supplied. I plan to address this in a separate commit.