Fix Account::PermissionsController#update
By duplicating the tests and implementation we've recently added for
`Users::PermissionsController#update`.

We discussed extracting a method to remove the duplication between the
two `build_user_update_params` methods but couldn't agree on where the
responsibility should live, so we've decided to leave it for now.

Note the change to `PunditHelpers#stub_policy` where we've updated it to
handle namespaced policies.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>