Stub all the policies used in the navigation items
The `NavigationItemsHelper#navigation_items` method calls the `#index?`
predicate method for a number of model classes. Although we haven't
actually needed these methods to be stubbed so far, I'm about to make
some changes which will need at least one of them. And it seems sensible
to me to stub all of them given the purpose of
`PunditHelpers#stub_policy_for_navigation_links`.