Use PunditHelpers#stub_policy in Users::InvitationResendsControllerTest
This removes some duplication and makes the test terser and more
consistent with other similar controller tests.

It also makes the stubbing a bit stricter which is why we now also have
to call `PunditHelpers#stub_policy_for_navigation_links` in tests which
render a page.

I'm not 100% convinced that the latter is a good idea, because it makes
the tests noisier & more brittle. However, making the tests do this
stubbing consistently is more important to me at the moment and in
itself will make it easier to change the approach in the future.