Use new AuthorisationPolicy in AuthorisationsController
Previously the actions in this controller were using predicate methods
on `ApiUserPolicy` to determine whether a user had permission to carry
out the relevant action. This meant that the same `ApiUserPolicy#new?`
predicate method was being used for both `ApiUsersController#new` and
`AuthorisationsController#new` actions. While it makes sense that the
logic is the same (superadmin access only) for both actions, it's not
very idiomatic and I found it quite confusing. Additionally I'm planning
to add some more actions to this controller which would exacerbate the
problem.

The authorisations referenced by this controller are actually instances
of `Doorkeeper::AccessToken` which is quite confusing. However, the
terminology is quite widespread across the codebase and so I've decided
to stick with it for now. This meant that I had to use the
`policy_class` option in the call to `Pundit::Authorization#authorize`
in `AuthorisationsController#authorize_authorisation` so that I could
name the policy class `AuthorisationPolicy` vs
`Doorkeeper::AccessTokenPolicy`.