Remove unused module from AuthorisationsController
None of the methods defined in `UserPermissionsControllerMethods` seem
to be used in the `AuthorisationsController` or its only view template,
`app/views/authorisations/new.html.erb`, so it should be safe to stop
including it.

Looking through the git history it's not immediately obvious to me that
this module was ever needed, but I might have missed something.