Split out new page for revoking API user access tokens
This commit changes the "manage tokens for API user" page to link to a
new page where you can revoke a specific token for an API user. This is
a small step on the way to changing the "manage tokens for API user"
page to use the GOV.UK Design System.

There wasn't any design for this page, so I've just made it up. I
thought it was useful to include some details about the access token
that's about to be revoked and so I've used a `summary_list` component
for this purpose. I hope that makes sense!