Move API user manage tokens page to GOV.UK Design System
Trello: https://trello.com/c/75Jyg8zR

The main change to the page is that the access tokens for the API user
are now displayed as "Summary card" components rather than as rows in a
table. The "Summary card" component is documented here [1] as wrapper
around a `summary_list` component, but it doesn't currently exist in the
Rails component library, so I've implemented it using lower-level GOV.UK
Design System CSS classes.

Similarly, I've had to use lower-level CSS classes for the summary lists
inside each summary card, because there's no support for a custom link
(i.e. "Revoke") in the Rails component library version of
`summary_list`.

The "Add application token" button has moved to the top of the page as
per the design.

I've replaced the "danger" & "success" alerts with a single
`success_alert` component [1]. There is no real equivalent to the
"danger" alert in the GOV.UK Design System - the `error_alert`
component [2] isn't appropriate, because there's not actually an error
in this case - it's just trying to draw attention to the fact that the
access token code will only be visible temporarily - so using a single
`success_alert` made the most sense to me. I think the copy could be
improved, but I'm going to say that's out-of-scope for now.

In order to display the `success_alert` in the right place, I've
introduced a new `custom_alerts` `content_for` call in the
`admin_layout`.

I've replaced the functionality provided by `ZeroClipboard` (and the
`zeroclipboard-rails` gem) in this alert with the `copy_to_clipboard`
component [3]. Note that the `ZeroClipboard` behaviour wasn't working
for me (at least in Firefox), but the new component works fine.

Lastly I've had to tweak a few of the assertions in existing tests to
account for the changes in the markup.

[1]: https://design-system.service.gov.uk/components/summary-list/#summary-cards
[2]: https://components.publishing.service.gov.uk/component-guide/success_alert
[3]: https://components.publishing.service.gov.uk/component-guide/error_alert
[4]: https://components.publishing.service.gov.uk/component-guide/copy_to_clipboard