Reorganize contexts in AuthorisationsControllerTest
This moves the action-related contexts up to the top-level and the
contexts related to the type of the current user down to the next level.
I think this is easier to reaad and it is more consistent with other
controller tests.

All the tests remain unchanged except for the one which was sending a
GET request to the `revoke` action. I've corrected this so that it now
sends a POST request which is specified by the route.

I plan to add more tests for the `revoke` action and so I've deleted the
comment, because it *is* in fact possible to test non-RESTful actions
from a controller test!