Improve Manage API Users integration test
The comment in the `assert_has_signin_permission_for` method was
incorrect. The presence of an application in the "Permissions" list is
determined by `UserPermissionsControllerMethods#visible_applications`,
which uses `User#authorisations` and not `User#application_permissions`.

I've renamed the original method to better describe its behaviour, and
I've added a new method to test that the user has the signin permission
for the app. Unfortunately there's no way in the user interface to know
whether an API User has the signin permission for an app so we have to
rely on `User#has_access_to?`.