Remove redundant zeroclipboard-rails gem
And associated JavaScript config.

This was originally added in this PR [1] and I can't see any other
remnants.

This seems like a good thing, because the functionality it was supposed
to provide wasn't working for me (at least in Firefox) and the GitHub
repo [2] has been archived since 2021!

We're now using the `copy_to_clipboard` component [3] instead.

[1]: https://github.com/alphagov/signon/pull/248
[2]: https://github.com/zeroclipboard/zeroclipboard-rails
[3]: https://components.publishing.service.gov.uk/component-guide/copy_to_clipboard