Simplify ApiUsersController#visible_applications
We use the ApiUserPolicy to prevent non-superadmin users from accessing
any of the actions in this controller.

The visible_applications method is only used in this controller: in
`index` via `ApiUsersHelper#application_list` and in
`manage_permissions` via `api_user_applications_and_permissions`, so we
can rely on our Pundit policy to ensure the current user is a
superadmin.