Convert SigninRequiredAuthorizationsController to design system
This controller was previously using the legacy design to render its
error message. This commit ensures that it uses the design system
instead.

I've decided to use an `error_alert` component here and keep the message
and description from the original devise-provided template which we
copied into our code in a previous commit.