Move call to layout to ApplicationController
We can move this call to `layout` to the ApplicationController to
remove duplication from the controllers that inherit from it. Note the
SigninRequiredAuthorizationsController retains a call to `layout` to
set the `admin_layout` as it doesn't inherit from
ApplicationController. The PasswordsController also has an explict
reference to `admin_layout` in a call to `render`. Later we will be
able to make `admin_layout` the default layout and remove those calls
too.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>