Remove layouts/application.html.erb
This layout is no longer used by the application so can be
removed. We've also removed the dependent legacy stylesheets, javascript and
the vendored `chosen` code as well as the shared
bootstrap_flash_messages partial and its helper since these are no
longer used elsewhere in the app.

Note that since 5b0f80d4 we are implementing the
same sanitization of tokens in URLs that was being performed in this
layout by the `sensitive_query_parameters?` block.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>