WIP: remove devise unlock route
TODO: check logit to make sure users are not requesting these urls

We noticed that these routes, provided by devise, were being rendered
using a devise template inside the new design-system layout. Rather
than convert them to use the design system we think we can remove them
altogether because our UserMailer#unlock_instructions email doesn't
include a link to unlock_url (this is different to the default devise
email for user unlocking[1]). It looks like[2] we've never allowed
users to unlock their own accounts, but that we did plan to allow that
at some point.

For now, rather than have a potential unstyled page in the
application, we'll remove these routes altogether. If we do decide to
allow unlocking in the future we'll need to make sure the
devise-provided views are either over-ridden or styled correctly.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>

[1] https://github.com/heartcombo/devise/blob/e2242a95f3bb2e68ec0e9a064238ff7af6429545/app/views/devise/mailer/unlock_instructions.html.erb#L7C4-L7C4
[2] 5c70330d5f1b4e