Configure date_formats
These date/time formats are currently provided by the
govuk_admin_template gem[1]. We want to remove that gem since we're
not using most of what it provides now we've moved to the new design
system. I've copied the formats from the gem into our date_formats
initializer so we can continue using them after the gem is removed.

[1] https://github.com/alphagov/govuk_admin_template/blob/5fc0de2dd570e12b5d816261cbe4a4bc18752845/lib/govuk_admin_template/engine.rb#L16