Extract UserUpdatePermissionBuilder
This commit moves the identical build_user_update_params method from
each of the three controllers into a class. I think this is an
improvement as it ensures that the business logic is defined in a
single place and can't drift out of sync.

The implementation of this method is currently covered by the
controller tests but in a subsequent commit I'll introduce some unit
tests for this class to help reduce the duplication in the controller
tests.