Use `Roles::<class>.require_2sv?`
The 3 lines changed in this commit were all checking that the role was
anything other than "normal" but in 3 different ways. Using the recently
introduced `Roles::<class>.require_2sv?` allows us to remove this
duplication.

Note that we have to fallback to `false` in
`InvitationsController#invitee_requires_2sv` to prevent a database
constraint violation attempting to insert `null` into
`User#require_2sv`.