Use table component to render users/application/index tables
This helps remove some of the duplication in the markup and ensures
that any changes that are made to the table component upstream are
reflected in signon when we update the govuk_publishing_components
gem.

Note that the table component does not support setting the width of
the columns or aligning the cells to the right (unless we use the
"numeric" format argument in `rows` which doesn't seem appropriate
semantically here). I'll address that in a subsequent commit.