Adopt whitehall-style "govuk-table--with-actions" pattern
The `govuk_publishing_components` `table` component does not support
aligning columns (without resorting to setting the format to `numeric`
which doesn't feel semantically correct). To work around this
whitehall wraps these tables in a div with class
`govuk-table--with-actions` and aligns the final cell in each row with
a CSS override[1].

After discussion with Calum to determine that it is important to
visually align these actions to the right, I've adopted something
similar here. Since our final actions column can have a button-styled
link as well as a text one, I've decided to align them to the right
with a small gap between using a flex container.

When this lands I'll raise the fact that this is a pattern that is now
being used in a few places across `alphagov` repos and see if it's
something that might be considered as a change to push up to
`govuk_publishing_components`. That would allow us and whitehall to
remove this potentially brittle CSS override.

[1] https://github.com/search?q=org%3Aalphagov%20govuk-table--with-actions&type=code