Reorder roles in `Roles::Superadmin.manageable_roles`
To match the order of the roles in the other
`Roles::<class>.manageable_roles` methods. Making the implementations
consistent will make it easier to remove duplication.

This also has the side effect of consistently ordering the roles in the
"Roles" filter on the /users page. Previously the "Roles" filter was
ordered from Superadmin -> Normal for Superadmin users, and from Normal
-> Superadmin for all other users.