Return empty string rather than nil from ApplicationTableHelper
The table component renders the string "Not set" if a row is provided
with `nil` for one of the `text` values[1]. This isn't very useful in
the context of the columns of these tables that have a visually hidden
header and contain "actions" (e.g. "Remove access"). Returning an
empty string instead ensures that the cell is rendered blank in this case.

[1] https://github.com/alphagov/govuk_publishing_components/blob/ba139b8a2e77a81c93484173101238ca1c40d001/spec/components/table_spec.rb#L47