Improve User validation messages
Updated content provided by Etain.

The new error messages include the attribute name within the string so
we have to use `#message` and `#message_for` (instead of `#full_message`
and `#full_messages_for` respectively) in order to avoid the attribute
name being prepended to the message.

We show these new error messages in:

- devise/invitations/new.html.erb
  - creating a new user
- api_users/new.html.erb
  - creating a new api user
- account/emails/edit.html.erb
  - changing your own email (Note there's not currently a way to change
    your own name)
- users/emails/edit.html.erb
  - changing a user or api user's email
- users/names/edit.html.erb
  - changing a user or api user's name