Update confirmation_instructions email
Updated content provided by Etain.

This is the email that's sent to the user's new email address when they
request to change their password.

Note that I have to manually remove the closing parenthesis from the
confirmation URL in the email_change integration test. Capybara's
`find_link` method was including the closing parenthesis which meant
that clicking on the link was failing.