Replace hardcoded tables with the component
We want to be consistent in the use of the custom table component in the
signon app.

Converting the two hardcoded tables to component-built tables results in
losing the ability to set a header style for a column. We think this is
worth it, for the benefit of using the same table component across the
app.