Improve error message for `User#reason_for_2sv_exemption`
So that it makes sense when appended onto the field name, and so that it
uses `#humanize` to render a human friendly version of the role name.

From:

"Reason for 2sv exemption admin users cannot be exempted from 2SV.
Remove the user's exemption to change their role."

To:

"Reason for 2sv exemption cannot be blank for Admin users. Remove
the user's exemption to change their role."