Fix assertion in permissions_promoter_test
I'm not sure why but this assertion wasn't working as expected. I
noticed it when I made a change to the block such that it was raising an
exception but the exception was being silently ignored and the test
continued to pass.

You can reproduce the problem with this simple test case:

```
require "test_helper"

class FooTest < ActiveSupport::TestCase
  should "fail if block raises an exception" do
    assert [1, 2].all? do |i|
      raise "error"
    end
  end
end
```